backend/requirements.txt
Django>=3.2,<5.0
Pillow>=10.1.0,<11.3.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.1,<5.4.0
psycopg2-binary>=2.9.9,<2.10.0
dj-database-url>=2.1.0,<2.2.0
django-debug-toolbar>=4.2.0,<4.3.0
django-redis>=5.4.0,<5.5.0
django-cors-headers>=3.7.0,<4.3.1
python-decouple==3.8
django-ckeditor
django-jazzmin
whitenoise
django-user-agents
PyJWT
psycopg2-binary
